+++ backend/garpix_user/serializers/auth_token_serializer.py
@@ -23,15 +23,7 @@ class AuthTokenSerializer(serializers.Serializer):
             request = self.context.get('request')
             user = authenticate(request=request,
                                 username=username, password=password)
-            current_user_session = UserSession.get_from_request(request)
-            if current_user_session:
-                if (user_user_session := UserSession.objects.filter(
-                        user=user).first()) and user_user_session != current_user_session:
-                    current_user_session.delete()
-                else:
-                    current_user_session.user = user
-                    current_user_session.recognized = UserSession.UserState.REGISTERED
-                    current_user_session.save()
+            user.set_user_session(request)
             # The authenticate call simply returns None for is_active=False
             # users. (Assuming the default ModelBackend authentication
             # backend.)