+++ backend/garpix_user/views/login_views.py
@@ -30,15 +30,7 @@ class LoginView(FormView):
         username = data.get('username')
         password = data.get('password')
         user = authenticate(request, username=username, password=password)
-        current_user_session = UserSession.get_from_request(request)
-        if current_user_session:
-            if (user_user_session := UserSession.objects.filter(
-                    user=user).first()) and user_user_session != current_user_session:
-                current_user_session.delete()
-            else:
-                current_user_session.user = user
-                current_user_session.recognized = UserSession.UserState.REGISTERED
-                current_user_session.save()
+        user.set_user_session(request)
         login(request, user)
         if self.request.accepts('text/html'):
             return redirect(request.GET.get('next', '/'))