+++ backend/garpix_user/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Restore password bugs fixed
 - UserSession bugs fixed
 - Registration bugs fixed
+- Method `set_user_session` added to `User` model (see `Readme.md`)
 
 ### 3.1.0 (18.01.2023)